requirements.txt
neo4j==5.15.0
anthropic>=0.18.0
tqdm>=4.65.0
python-dotenv>=1.0.0
requests>=2.25.0

# Django web interface
django>=4.2.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0